Refactor code to index sub module
